mudando para api carrinho botoes e estilização

+++ src/app/pages/cart-dialog/cart-dialog.component.ts
@@ -21,8 +21,22 @@ export class CartDialogComponent {
     return this.cartItems.reduce((sum, item) => sum + item.price * item.quantity, 0);
   }
 
+  addItem(id: number) {
+    const item = this.cartItems.find(p => p.id === id);
+    if (item) {
+      this.cartService.addToCart(item);
+      this.cartItems = this.cartService.getCartItems();
+    }
+  }
+
   removeItem(id: number) {
-    this.cartService.removeFromCart(id);
+    const item = this.cartItems.find(p => p.id === id);
+    if (item && item.quantity > 1) {
+      item.quantity -= 1;
+      this.cartService.updateCart(this.cartItems);
+    } else {
+      this.cartService.removeFromCart(id);
+    }
     this.cartItems = this.cartService.getCartItems();
   }
 
@@ -37,7 +51,6 @@ export class CartDialogComponent {
     finalizarCompra() {
       this.orderService.addPedido([...this.cartItems]);
       this.clear();
-      this.dialogRef.close();
     }
   }
